requirements.txt
pytest
pylint
numpy
pandas
matplotlib
seaborn
scikit-learn
scipy
streamlit
sklearn
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
Pillow

# API
#fastapi
#pytz
#uvicorn
